backend/event_collector/requirements.txt
nostr-sdk==0.36.0
celery==5.3.6
redis==5.0.1
loguru==0.7.2
pyyaml==6.0.2